# my-blog-end/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/common/reply/query": {
            "post": {
                "description": "博客/回复的回复列表",
                "tags": [
                    "reply"
                ],
                "summary": "回复列表",
                "parameters": [
                    {
                        "description": "QueryReplyBody",
                        "name": "QueryReplyBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.QueryReplyBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.PageResult-controller_ReplyVo"
                        }
                    }
                }
            }
        },
...
        "/getOtherUserInfo/:userId": {
            "get": {
                "description": "获取其他用户的信息",
                "tags": [
                    "user"
                ],
                "summary": "获取其他用户的信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.UserVo"
                        }
                    }
                }
            }
        },
        "/getOtherUserStats/:userId": {
            "get": {
                "description": "获取其他用户的统计信息",
                "tags": [
                    "user"
                ],
                "summary": "获取其他用户的统计信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.StatVo"
                        }
                    }
                }
            }
        },
...
        "/reply/add": {
            "post": {
                "description": "回复博客/回复",
                "tags": [
                    "reply"
                ],
                "summary": "回复",
                "parameters": [
                    {
                        "description": "ReqReply",
                        "name": "ReqReply",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqReply"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/reply/getMyReply": {
            "post": {
                "description": "我的所有回复列表",
                "tags": [
                    "reply"
                ],
                "summary": "我的所有回复列表",
                "parameters": [
                    {
                        "description": "PageParams",
                        "name": "PageParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.PageParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.PageResult-controller_ReplyVo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.PageParams": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
...
        "controller.PageResult-controller_ReplyVo": {
            "type": "object",
            "properties": {
                "all": {
                    "type": "integer"
                },
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.ReplyVo"
                    }
                },
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "controller.QueryReplyBody": {
            "type": "object",
            "required": [
                "parentId"
            ],
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "parentId": {
                    "type": "integer"
                }
            }
        },
        "controller.ReplyVo": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "authId": {
                    "type": "integer"
                },
                "authName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdTime": {
                    "type": "string"
                },
                "likeFlag": {
                    "type": "boolean"
                },
                "likeNum": {
                    "type": "integer"
                },
                "parentId": {
                    "type": "integer"
                },
                "replyNum": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "toAuthId": {
                    "type": "integer"
                },
                "toAuthName": {
                    "type": "string"
                }
            }
        },
...
        "controller.ReqReply": {
            "type": "object",
            "required": [
                "content",
                "parentId",
                "toAuthId"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "toAuthId": {
                    "type": "integer"
                }
            }
        },
...
        "controller.StatVo": {
            "type": "object",
            "properties": {
                "articleNum": {
                    "description": "发布博客数量、点赞量、被回复量、被点赞量",
                    "type": "integer"
                },
                "beLikedNum": {
                    "type": "integer"
                },
                "beRepliedNum": {
                    "type": "integer"
                },
                "likeNum": {
                    "type": "integer"
                }
            }
        },
        "controller.UserVo": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}